+++ data-transfer/src/main/java/it/eng/datatransfer/service/DataTransferService.java
@@ -5,6 +5,7 @@ import java.util.UUID;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
+import it.eng.datatransfer.event.TransferProcessChangeEvent;
 import it.eng.datatransfer.exceptions.AgreementNotFoundException;
 import it.eng.datatransfer.exceptions.TransferProcessExistsException;
 import it.eng.datatransfer.exceptions.TransferProcessInvalidStateException;
@@ -15,6 +16,7 @@ import it.eng.datatransfer.model.TransferProcess;
 import it.eng.datatransfer.model.TransferRequestMessage;
 import it.eng.datatransfer.model.TransferStartMessage;
 import it.eng.datatransfer.model.TransferState;
+import it.eng.datatransfer.model.TransferSuspensionMessage;
 import it.eng.datatransfer.properties.DataTransferProperties;
 import it.eng.datatransfer.repository.TransferProcessRepository;
 import it.eng.tools.client.rest.OkHttpRestClient;
@@ -113,7 +115,10 @@ public class DataTransferService {
 
 		TransferProcess transferProcessStarted = transferProcessRequested.copyWithNewTransferState(TransferState.STARTED);
 		transferProcessRepository.save(transferProcessStarted);
-		publisher.publishEvent(transferProcessStarted);
+		publisher.publishEvent(TransferProcessChangeEvent.Builder.newInstance()
+				.oldTransferProcess(transferProcessRequested)
+				.newTransferProcess(transferProcessStarted)
+				.build());
 		return transferProcessStarted;
 	}
 	
@@ -135,11 +140,32 @@ public class DataTransferService {
 
 		TransferProcess transferProcessCompleted = transferProcessStarted.copyWithNewTransferState(TransferState.COMPLETED);
 		transferProcessRepository.save(transferProcessCompleted);
-		publisher.publishEvent(transferProcessCompleted);
+		publisher.publishEvent(TransferProcessChangeEvent.Builder.newInstance()
+				.oldTransferProcess(transferProcessStarted)
+				.newTransferProcess(transferProcessCompleted)
+				.build());
 		return transferProcessCompleted;
-
 	}
 
+	public TransferProcess suspendDataTransfer(TransferSuspensionMessage transferSuspensionMessage, String consumerPid,
+			String providerPid) {
+		String consumerPidFinal = consumerPid == null ? transferSuspensionMessage.getConsumerPid() : consumerPid;
+		String providerPidFinal = providerPid == null ? transferSuspensionMessage.getProviderPid() : providerPid;
+		log.debug("Suspending data transfer for consumerPid {} and providerPid {}", consumerPidFinal, providerPidFinal);
+
+		TransferProcess transferProcessStarted = findTransferProcess(consumerPidFinal, providerPidFinal);
+		stateTransitionCheck(transferProcessStarted, TransferState.SUSPENDED);
+
+		TransferProcess transferProcessSuspended = transferProcessStarted.copyWithNewTransferState(TransferState.SUSPENDED);
+		transferProcessRepository.save(transferProcessSuspended);
+		publisher.publishEvent(TransferProcessChangeEvent.Builder.newInstance()
+				.oldTransferProcess(transferProcessStarted)
+				.newTransferProcess(transferProcessSuspended)
+				.build());
+		publisher.publishEvent(transferSuspensionMessage);
+		return transferProcessSuspended;
+	}
+	
 	private TransferProcess findTransferProcess(String consumerPid, String providerPidFinal) {
 		TransferProcess transferProcessRequested = transferProcessRepository.findByConsumerPidAndProviderPid(consumerPid, providerPidFinal)
 			.orElseThrow(() -> new TransferProcessNotFoundException("Transfer process for consumerPid " + consumerPid